requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
sqlalchemy==2.0.23
sqlite3-worker==0.1.9
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
msal==1.25.0
openpyxl==3.1.2
plotly==5.17.0
streamlit-aggrid==0.3.4.post3
concurrent-futures==3.1.1
threading-utils==0.3
